requirements.txt
opencv-python
torch
torchvision
pandas
numpy
scikit-learn
Pillow
matplotlib
fontforge
svgwrite
potrace